requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.23.0
cryptography==3.4.6
Django==3.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.3.0
idna==2.10
Pillow==8.1.1
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
